refactor(video-conferencing-service): implement vonage service to reduce lines of code in vonage pro
SFO-64
