+++ services/video-conferencing-service/src/providers/vonage/vonage.provider.ts
@@ -1,6 +1,4 @@
-import {Provider, inject, config} from '@loopback/core';
-import {HttpErrors} from '@loopback/rest';
-import {VonageEnums} from '../../enums/video-chat.enum';
+import { Provider, service } from '@loopback/core';
 import {
   ArchiveResponse,
   ArchiveResponseList,
@@ -8,110 +6,48 @@ import {
 } from '../../types';
 import {
   VonageVideoChat,
-  VonageConfig,
   VonageMeetingOptions,
   VonageMeetingResponse,
   VonageSessionOptions,
   VonageS3TargetOptions,
   VonageAzureTargetOptions,
 } from './types';
-import OpenTok from 'opentok';
-import axios from 'axios';
-import {repository} from '@loopback/repository';
-import {AuditLogsRepository} from '../../repositories';
-import {VonageBindings} from './keys';
-import {promisify} from 'util';
+
+import { VonageService } from '../../services/vonage.service';
+import { AuditLogsRepository } from '../../repositories';
+import { repository } from '@loopback/repository';
 import moment from 'moment';
-import {sign} from 'jsonwebtoken';
+import { HttpErrors } from '@loopback/rest';
 
 export class VonageProvider implements Provider<VonageVideoChat> {
   constructor(
-    @inject(VonageBindings.config)
-    private readonly vonageConfig: VonageConfig,
+    @service(VonageService)
+    private readonly vonageService: VonageService,
     @repository(AuditLogsRepository)
     private readonly auditLogRepository: AuditLogsRepository,
-  ) {
-    const {apiKey, apiSecret} = vonageConfig;
-    if (!(apiKey && apiSecret)) {
-      throw new HttpErrors.BadRequest('Vonage API key or secret is not set');
-    }
-    this.VonageService = new OpenTok(apiKey, apiSecret);
-  }
-
-  VonageService: OpenTok;
-
+  ) { }
   value() {
     return {
       getMeetingLink: async (
         meetingOptions: VonageMeetingOptions,
       ): Promise<VonageMeetingResponse> => {
-        let mediaMode: VonageEnums.MediaMode = VonageEnums.MediaMode.Routed;
-        let archiveMode: VonageEnums.ArchiveMode =
-          VonageEnums.ArchiveMode.Manual;
-        let sessionId: string;
-
-        const {endToEndEncryption, enableArchiving} = meetingOptions;
-
-        if (endToEndEncryption && enableArchiving) {
-          throw new HttpErrors.BadRequest(
-            'End to end Encryption along with archiving is not possible',
-          );
-        } else if (endToEndEncryption) {
-          mediaMode = VonageEnums.MediaMode.Relayed;
-        } else if (enableArchiving) {
-          archiveMode = VonageEnums.ArchiveMode.Always;
-        } else {
-          // nothing to do.
-        }
-
-        const sessionCreationOptions = {
-          mediaMode,
-          archiveMode,
-        };
-
-        const createSession = () => {
-          return new Promise<OpenTok.Session>((resolve, reject) => {
-            this.VonageService.createSession(
-              sessionCreationOptions,
-              (err: Error | null, session: OpenTok.Session | undefined) => {
-                if (err) {
-                  reject(err);
-                }
-                if (!session) {
-                  throw new HttpErrors.InternalServerError(
-                    'Error creating session',
-                  );
-                } else {
-                  resolve(session);
-                }
-              },
-            );
-          });
-        };
-
         try {
-          const session = await createSession();
+          const response = await this.vonageService.getMeetingLink(meetingOptions);
           this.auditLogRepository.create({
             action: 'session',
             actionType: 'create-session',
-            before: sessionCreationOptions,
-            after: session,
+            before: meetingOptions,
+            after: response,
             actedAt: moment().format(),
-            actedEntity: session.sessionId,
           });
-          return {
-            mediaMode: sessionCreationOptions.mediaMode,
-            archiveMode: sessionCreationOptions.archiveMode,
-            sessionId: session.sessionId,
-          };
-        } catch (err) {
+          return response;
+        } catch (error) {
           this.auditLogRepository.create({
             action: 'session',
             actionType: 'create-session',
-            before: sessionCreationOptions,
-            after: err.stack,
+            before: meetingOptions,
+            after: { errorStack: error.stack },
             actedAt: moment().format(),
-            actedEntity: '',
           });
           throw new HttpErrors.InternalServerError('Error creating session');
         }
@@ -121,17 +57,7 @@ export class VonageProvider implements Provider<VonageVideoChat> {
         options: VonageSessionOptions,
       ): Promise<SessionResponse> => {
         try {
-          const {expireTime, role, data} = options;
-          const requestPayload: OpenTok.TokenOptions = {
-            expireTime: expireTime ? moment(expireTime).unix() : undefined,
-            role: role ?? undefined,
-            data: data ?? undefined,
-          };
-          const token = this.VonageService.generateToken(
-            sessionId,
-            requestPayload,
-          );
-
+          const response = await this.vonageService.getToken(sessionId, options);
           this.auditLogRepository.create({
             action: 'session',
             actionType: 'get-token',
@@ -140,15 +66,11 @@ export class VonageProvider implements Provider<VonageVideoChat> {
               ...options,
             },
             after: {
-              token,
+              token: response.token,
             },
             actedAt: moment().format(),
-            actedEntity: token,
           });
-          return {
-            sessionId,
-            token,
-          };
+          return response;
         } catch (error) {
           this.auditLogRepository.create({
             action: 'session',
@@ -167,51 +89,25 @@ export class VonageProvider implements Provider<VonageVideoChat> {
           );
         }
       },
-
       getArchives: async (
         archiveId: string | null,
       ): Promise<ArchiveResponse | ArchiveResponseList> => {
         try {
-          const getArchive = promisify(this.VonageService.getArchive);
-          const listArchives = promisify(this.VonageService.listArchives);
-          let archiveResult: ArchiveResponse | ArchiveResponseList;
-          if (archiveId) {
-            const archive = await getArchive(archiveId);
-            archiveResult = {
-              name: archive?.name,
-              sessionId: archive?.sessionId as string,
-              metaData: (archive ?? null) as object,
-            };
-          } else {
-            let archives = await listArchives({});
-            archives = archives?.length ? archives : [];
-            const items = [];
-            for (const archive of archives) {
-              items.push({
-                name: archive.name,
-                sessionId: archive.sessionId as string,
-                metaData: archive as object,
-              });
-            }
-            archiveResult = {
-              count: archives?.length ?? 0,
-              items,
-            };
-          }
+          const response = await this.vonageService.getArchives(archiveId);
           this.auditLogRepository.create({
             action: 'archive',
             actionType: archiveId ? 'getArchive' : 'getArchives',
-            before: archiveId ? {archiveId} : {},
-            after: archiveResult,
+            before: archiveId ? { archiveId } : {},
+            after: response,
             actedAt: moment().format(),
           });
-          return archiveResult;
+          return response;
         } catch (error) {
           this.auditLogRepository.create({
             action: 'archive',
             actionType: archiveId ? 'getArchive' : 'getArchives',
-            before: archiveId ? {archiveId} : {},
-            after: {errorStack: error.stack},
+            before: archiveId ? { archiveId } : {},
+            after: { errorStack: error.stack },
             actedAt: moment().format(),
           });
           throw new HttpErrors.InternalServerError(
@@ -219,26 +115,16 @@ export class VonageProvider implements Provider<VonageVideoChat> {
           );
         }
       },
-      // TODO: startArchive and stopArchive needs to be uncommented/modified later
-      // startArchive: async (sessionId: string, archiveOptions: ArchiveOptions): Promise<Archive | undefined> => {
-      //   const startArchive = promisify(this.VonageService.startArchive);
-      //   return startArchive(sessionId, archiveOptions);
-      // },
-      // stopArchive: async (archiveId: string): Promise <Archive | undefined> => {
-      //   const stopArchive = promisify(this.VonageService.stopArchive);
-      //   return stopArchive(archiveId);
-      // },
       deleteArchive: async (archiveId: string) => {
         try {
-          const deleteArchive = promisify(this.VonageService.deleteArchive);
-          await deleteArchive(archiveId);
+          await this.vonageService.deleteArchive(archiveId);
           this.auditLogRepository.create({
             action: 'archive',
             actionType: 'delete-archive',
             before: {
               archiveId,
             },
-            after: {response: 'Archive Deletion Successful!'},
+            after: { response: 'Archive Deletion Successful!' },
             actedAt: moment().format(),
           });
         } catch (error) {
@@ -248,7 +134,7 @@ export class VonageProvider implements Provider<VonageVideoChat> {
             before: {
               archiveId,
             },
-            after: {errorStack: error.stack},
+            after: { errorStack: error.stack },
             actedAt: moment().format(),
           });
           throw new HttpErrors.InternalServerError(
@@ -262,75 +148,20 @@ export class VonageProvider implements Provider<VonageVideoChat> {
         const auditLogAction = 'archive';
         const auditLogActionType = 'set-storage-target';
         try {
-          const {apiKey, apiSecret} = this.vonageConfig;
-          const ttl = 200;
-          const jwtPayload = {
-            iss: apiKey,
-            ist: 'project',
-            iat: moment().unix(),
-            exp: moment()
-              .add(ttl, 'seconds')
-              .unix(),
-          };
-
-          const token = sign(jwtPayload, apiSecret);
-          let type = '';
-          const credentials = {};
-          const {
-            accessKey,
-            secretKey,
-            bucket,
-            endpoint,
-          } = storageConfig as VonageS3TargetOptions;
-          const {
-            accountName,
-            accountKey,
-            container,
-            domain,
-          } = storageConfig as VonageAzureTargetOptions;
-          if (accessKey && secretKey && bucket) {
-            type = 'S3';
-            Object.assign(credentials, {
-              accessKey,
-              secretKey,
-              bucket,
-              endpoint,
-            });
-          }
-          if (accountName && accountKey && container) {
-            type = 'Azure';
-            Object.assign(credentials, {
-              accountName,
-              accountKey,
-              container,
-              domain,
-            });
-          }
-          await axios({
-            url: `https://api.opentok.com/v2/project/${this.vonageConfig.apiKey}/archive/storage`,
-            method: 'put',
-            data: {
-              type,
-              config: credentials,
-              fallback: storageConfig.fallback,
-            },
-            headers: {
-              'X-OPENTOK-AUTH': token,
-            },
-          });
+          await this.vonageService.setUploadTarget(storageConfig);
           this.auditLogRepository.create({
             action: auditLogAction,
             actionType: auditLogActionType,
-            before: config,
-            after: {response: 'Storage Target Success'},
+            before: storageConfig,
+            after: { response: 'Storage Target Success' },
             actedAt: moment().format(),
           });
         } catch (error) {
           this.auditLogRepository.create({
             action: auditLogAction,
             actionType: auditLogActionType,
-            before: config,
-            after: {errorStack: error.stack},
+            before: storageConfig,
+            after: { errorStack: error.stack },
             actedAt: moment().format(),
           });
           throw new HttpErrors.InternalServerError(
@@ -338,51 +169,6 @@ export class VonageProvider implements Provider<VonageVideoChat> {
           );
         }
       },
-      // TO-DO: will do modifications later
-      // deleteUploadTarget: async (): Promise<void> => {
-      //   const auditLogAction = 'archive';
-      //   const auditLogActionType = 'set-storage-target';
-      //   try {
-      //     const {apiKey, apiSecret} = this.vonageConfig;
-      //     const ttl = 200;
-      //     const jwtPayload = {
-      //       iss: apiKey,
-      //       ist: 'project',
-      //       iat: moment().unix(),
-      //       exp: moment()
-      //         .add(ttl, 'seconds')
-      //         .unix(),
-      //     };
-      //     const token = sign(jwtPayload, apiSecret);
-      //     await axios({
-      //       url: `https://api.opentok.com/v2/project/${process.env.TOKBOX_API_KEY}/archive/storage`,
-      //       method: 'delete',
-      //       headers: {
-      //         'X-OPENTOK-AUTH': token,
-      //       },
-      //     });
-      //     this.auditLogRepository.create({
-      //       action: auditLogAction,
-      //       actionType: auditLogActionType,
-      //       before: config,
-      //       after: {
-      //         response: 'successfully removed storage target from s3/azure ',
-      //       },
-      //       actedAt: moment().format(),
-      //     });
-      //   } catch (error) {
-      //     this.auditLogRepository.create({
-      //       action: auditLogAction,
-      //       actionType: auditLogActionType,
-      //       before: {},
-      //       after: {errorStack: error.stack},
-      //       actedAt: moment().format(),
-      //     });
-      //     throw new HttpErrors.InternalServerError(
-      //       'Error occured while removing s3/azure storage target',
-      //     );
-      //   }
-      // },
-    };
+    }
   }
 }
